FlaskProject_MQ/main_app/requirements.txt
flask==2.0.1
flask-cors==3.0.10
openai==1.12.0
python-dotenv==1.0.0
selenium
html2text
beautifulsoup4
requests
pika